Dynamic blog
Added a blog.json to my github repo, so now I can dynamically add blog posts.
Id a post is labelled as "featured" it has a random chance of being added to the featured article section. All articles (including featured) are put in the bottom section.

I have added an improved animation to the bottom section, which will make it look nice for articles with longer titles. The animation now slides up revealing some of the paragraph.